requirements.txt
python-telegram-bot[job-queue]
pandas
openpyxl
python-docx
pytz
python-dotenv